RenderArea/ updating classes and parser
I don't want to break anything so I made a new branch.

Created the render area class but in order for it to all work have a lot of updating to do in classes and parser so also was working on that. Not done yet but I think enough is there to show gist of it all