RenderArea can now Save()
calls every shape's serializer, verified to match original shapes.txt

PLEASE NOTE: since we can now edit the shapes.txt, there is no longer a guarantee that it is the original from Canvas! I've made a copy of the original one and put it in the root directory of the repo, for reference. This file ("shapes-original.txt") should not be edited.